Enable non-error-path Collector TCK tests
