--- vertx-db2-client/src/test/java/io/vertx/db2client/tck/DB2PeparedQueryCachedTest.java
@@ -1,36 +1,0 @@
-package io.vertx.db2client.tck;
-
-import io.vertx.db2client.junit.DB2Resource;
-import io.vertx.ext.unit.TestContext;
-import io.vertx.ext.unit.junit.VertxUnitRunner;
-import io.vertx.sqlclient.tck.PreparedQueryCachedTestBase;
-import org.junit.ClassRule;
-import org.junit.Ignore;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-@Ignore // TODO @AGG get this TCK test passing
-@RunWith(VertxUnitRunner.class)
-public class DB2PeparedQueryCachedTest extends PreparedQueryCachedTestBase {
-  @ClassRule
-  public static DB2Resource rule = DB2Resource.SHARED_INSTANCE;
-
-  @Override
-  protected void initConnector() {
-    options = rule.options();
-    connector = ClientConfig.CONNECT.connect(vertx, options);
-  }
-
-  @Override
-  protected String statement(String... parts) {
-    return String.join("?", parts);
-  }
-
-  @Test
-  @Ignore
-  @Override
-  public void testPreparedQueryParamCoercionTypeError(TestContext ctx) {
-    // Does not pass, we can't achieve this feature on MySQL for now, see io.vertx.mysqlclient.impl.codec.MySQLParamDesc#prepare for reasons.
-    super.testPreparedQueryParamCoercionTypeError(ctx);
-  }
-}
